+++ contracts/sdk/gummyroll/types/index.ts
@@ -1,0 +1,1 @@
+export { Gummyroll } from "../../../target/types/gummyroll";
